% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weo_get_latest_publication.R
\name{weo_get_latest_publication}
\alias{weo_get_latest_publication}
\title{Get Latest WEO Publication from IMF Website}
\usage{
weo_get_latest_publication(quiet = TRUE)
}
\arguments{
\item{quiet}{A logical indicating whether to print download information.
Defaults to TRUE.}
}
\value{
A list with year and release
}
\description{
Determines the latest available WEO publication based on the current date.
}
