% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imfweo-package.R
\docType{package}
\name{imfweo-package}
\alias{imfweo}
\alias{imfweo-package}
\title{imfweo: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Teal Emery \email{lte@tealinsights.com}

}
